core-backend/requirements.txt
# AI / Agent / Embedding / LLM stack
crewai==0.130.0
crewai-tools==0.47.1
embedchain==0.1.128
langchain-openai==0.2.1
openai==1.68.2
litellm==1.72.0
tiktoken>=0.8.0,<0.9.0


# The rest dependencies you had before (fastapi, sqlalchemy etc) remain same or adjusted if needed.

# Core FastAPI dependencies
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic==2.9.0
pydantic-settings==2.5.0
python-dotenv==1.0.0


# The rest:
sqlalchemy==2.0.35
alembic==1.13.1
psycopg2-binary==2.9.7
asyncpg==0.29.0
redis==5.0.1

python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
bcrypt==4.0.1

aiofiles==24.1.0
python-magic==0.4.27
pypdf==5.0.0
python-docx==1.1.0
openpyxl==3.1.5
pandas==2.2.2

structlog==24.1.0
sentry-sdk[fastapi]==2.13.0

pytest==8.3.0
pytest-asyncio==0.23.0
httpx==0.27.0
pytest-cov==5.0.0
black==24.8.0
isort==5.13.0
flake8==7.1.0
mypy==1.11.0

mkdocs==1.6.0
mkdocs-material==9.5.0

click==8.1.8
rich==13.7.0
typer==0.12.0
numpy==1.26.4

pillow==10.3.0
protobuf==5.27.0
oauthlib==3.2.2
jsonschema==4.22.0
bleach==6.1.0
email-validator==2.1.0

opentelemetry-api>=1.30.0
opentelemetry-exporter-otlp-proto-common>=1.30.0
opentelemetry-exporter-otlp-proto-grpc>=1.30.0
opentelemetry-exporter-otlp-proto-http>=1.30.0
opentelemetry-instrumentation>=0.51b0
opentelemetry-instrumentation-asgi>=0.51b0
opentelemetry-instrumentation-fastapi>=0.51b0
opentelemetry-proto>=1.30.0
opentelemetry-sdk>=1.30.0
opentelemetry-semantic-conventions>=0.51b0
opentelemetry-util-http>=0.51b0

faker==25.0.0
factory-boy==3.3.0
